+++ src/main/java/Restaurant/Restaurant/Dish/singleDish/service/DishService.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 @Service
 public interface DishService {
     List<Dish> getAll();
-    Dish addDish(String nazwa, float cena);
+    Dish addDish(String nazwa, Integer cena);
     Dish getByName(String Dishname);
-    void editDish(Long id, String name,float price);
+    void editDish(Long id, String name,Integer price);
     void removeDish(Long id);
     Optional<Dish> getById(Long id);
     boolean isNameUsed(String name);
